add constant file and refactor code
